Switched to 'react-hot-loader' restructured application a bit
