+++ .babelrc
@@ -6,13 +6,13 @@
   ],
   "presets": [
     ["es2015", { "modules": false }],
-    "react",
-    "stage-0"
+    "stage-0",
+    "react"
   ],
   "env": {
     "development": {
-      "presets": [
-        "react-hmre"
+      "plugins": [
+        "react-hot-loader/babel"
       ]
     }
   }
