+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-redux-webpack2-boilerplate",
-  "version": "0.1.7",
+  "version": "0.1.8",
   "private": false,
   "license": "MIT",
   "main": "index.js",
@@ -26,7 +26,6 @@
     "babel-plugin-transform-runtime": "^6.6.0",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.5.0",
-    "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-0": "^6.16.0",
     "babel-runtime": "^6.6.1",
     "css-loader": "0.14.5",
@@ -41,6 +40,7 @@
     "node-sass": "^3.13.0",
     "postcss-loader": "^1.1.1",
     "prepush": "^3.1.11",
+    "react-hot-loader": "^3.0.0-beta.7",
     "redux-logger": "^2.7.4",
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.0",
@@ -69,7 +69,7 @@
   "description": "Starter boilerplate for React and Redux, using Webpack 2",
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@github.com/Stanko/react-redux-webpack2-boilerplate.git"
+    "url": "git+ssh://git@github.com/workco/marvin.git"
   },
   "keywords": [
     "react",
@@ -79,9 +79,9 @@
   ],
   "author": "Stanko",
   "bugs": {
-    "url": "https://github.com/Stanko/react-redux-webpack2-boilerplate/issues"
+    "url": "https://github.com/workco/marvin/issues"
   },
-  "homepage": "https://github.com/Stanko/react-redux-webpack2-boilerplate#readme",
+  "homepage": "https://github.com/workco/marvin#readme",
   "prepush": [
     "npm run lint-break-on-errors --silent"
   ]
