+++ source/js/config/routes.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { Route, Switch } from 'react-router-dom';
+
+import Dashboard from 'views/Dashboard';
+import About from 'views/About';
+import NotFound from 'views/NotFound';
+
+const publicPath = '/';
+
+export const routeCodes = {
+  DASHBOARD: publicPath,
+  ABOUT: `${ publicPath }about`,
+};
+
+export default () => (
+  <Switch>
+    <Route exact path={ publicPath } component={ Dashboard } />
+    <Route path={ routeCodes.ABOUT } component={ About } />
+    <Route path='*' component={ NotFound } />
+  </Switch>
+);
