+++ source/js/views/App/index.jsx
@@ -1,18 +1,8 @@
 import React, { Component } from 'react';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import Routes from 'config/routes';
 import PropTypes from 'prop-types';
 
-import Dashboard from 'views/Dashboard';
-import About from 'views/About';
-import NotFound from 'views/NotFound';
 import Menu from 'components/Global/Menu';
-
-const publicPath = '/';
-
-export const routeCodes = {
-  DASHBOARD: publicPath,
-  ABOUT: `${ publicPath }about`,
-};
 
 export default class App extends Component {
   static propTypes = {
@@ -21,18 +11,12 @@ export default class App extends Component {
 
   render() {
     return (
-      <BrowserRouter>
-        <div className='App'>
-          <Menu />
-          <div className='Page'>
-            <Switch>
-              <Route exact path={ publicPath } component={ Dashboard } />
-              <Route path={ routeCodes.ABOUT } component={ About } />
-              <Route path='*' component={ NotFound } />
-            </Switch>
-          </div>
+      <div className='App'>
+        <Menu />
+        <div className='Page'>
+          <Routes />
         </div>
-      </BrowserRouter>
+      </div>
     );
   }
 }
