+++ source/js/views/Client/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { BrowserRouter } from 'react-router-dom';
+import App from 'views/App';
+
+export default () => (
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
+);
